KPME-2358: payout leave blocks being deleted by accrual service. KPME-2357: transfer leave blocks being deleted by accrual service.

+++ src/main/java/org/kuali/hr/lm/accrual/service/AccrualServiceImpl.java
@@ -370,6 +370,8 @@ public class AccrualServiceImpl implements AccrualService {
 					sstoUsageList.add(lb);
 				}
 			} else {
+				if(!(StringUtils.equals(lb.getLeaveBlockType(),LMConstants.LEAVE_BLOCK_TYPE.BALANCE_TRANSFER) ||
+						StringUtils.equals(lb.getLeaveBlockType(),LMConstants.LEAVE_BLOCK_TYPE.LEAVE_PAYOUT)))
 				TkServiceLocator.getLeaveBlockService().deleteLeaveBlock(lb.getLmLeaveBlockId(), runAsPrincipalId);
 			}
 		}
